feat: implement splash screen, decimal inventory support, and robust image uploading and fetching handling fix: enhanced the api response to parse nn-strng messages and handle business logic errors

+++ lib/features/low_stock/presentation/providers/low_stock_provider.dart
@@ -39,7 +39,7 @@ class LowStockProducts extends _$LowStockProducts {
   /// update a product's quantity
   Future<void> updateProductQuantity(
     ProductModel product,
-    int newQuantity,
+    double newQuantity,
   ) async {
     final repository = ref.read(productRepositoryProvider);
